// src/types/CsSelections.ts
// src/types/CsSelections.ts

export interface CsSelections {
    mathCourses: string[];
    scienceCourses: string[];
    introEngineeringCourse: string[];
    liberalArtsCourses: string[];
    csCoreCourses: string[];
    csDepthCourses: string[];
    csProjectCourse: string[];
    technicalElectives: string[];
    openElectives: string[];
    computersAndEthicsCourse: string[];
    writingComponentCourse: string[];
  }
  